Reworked mixing and buffer management. Still not quite there yet.
Now we use an SDL_AudioStream to manage resampling on the fly, so we don't
get artifacts between buffers in the queue. Querying buffer state is more
correct now too.

This broke several things (we need to resample single AL_BUFFER objects too
but don't, sample count math is wrong in the mixer, race conditions, etc), so
we have some work to do still.