Added tests/testposition.c for simple GUI for testing spatial rendering.
This has a single source (blue square), looked upon from above (so it moves
on the X and Z axes, and Y is always 0.0f). The listener is looking north,
facing up towards the camera (green square). You can drag either around as
the source loops a .wav of your choice.

This could be fancier than it is, but it's good enough for now.