NEON stereo buffer mixing code gets slight benefit from more unrolling.
We have 16 128-bit NEON registers, so we could afford to do it here, unlike
SSE1 which has only 8 registers. Since we don't need to unzip the stereo
samples, we had registers to spare, but the mono buffer mixer doesn't.

This isn't earth-shatteringly faster, but it did cause a small positive win.
Mostly I imagine I'm memory bound at this point, on the Raspberry Pi that I
was testing on, so I might as well let the compiler have the extra registers
to play with.