Wrap all the API entry points in a mutex.
This serializes access to the API, but multi-thread access to OpenAL is
probably uncommon in general, and the mixer thread does not ever hold this
lock. Since uncontended mutexes are generally pretty fast in modern times,
this shouldn't be a big deal but will make the code both more simple and less
fragile.