Don't assert in the mixer if source type is AL_UNDETERMINED.
Just remove it from the playlist.

This can happen due to a race condition that needs to be fixed, but it's
also legal to just alSourcePlay() a source with its AL_BUFFER set to zero,
so you can get into this state by legal means too, so handle it properly.