mixer: Initial shot at surround sound support!
Tested with quad, 4.1, 5.1, 7.1, 2.1, stereo and mono. I don't have a system
that can test 6.1, but I assume it'll work (or an array of numbers needs a
little tweaking).

Added mono output, too, for completeness. I doubt anyone will ever use it, so
I didn't bother filling in the obvious SIMD code for it.

Fixes #33.