Finished C99'ifying the codebase.
Mostly this is moving variables from the top of scope to a more meaningful
location. A few other tweaks and fixes.

This is always scary because it might introduce subtle bugs, but since so
much has changed about this code in the last week, this is as good a time as
any to accidentally break things.